fix #7: nothing is printed for google.com
This commit removes manual parsing done in TlsInspector because the
parsing sometimes failed to parse certificates from google.com. Instead
the code now gets the certificates by calling the peer_connections
method of the TLS connection. This not only fixes the bug but also
reduces the code we have to maintain.